Cache blobs
When summarizing an OCI we will now cache all the blobs inside it on
disk. This allows faster access to the blobs, because we won't have to
decompress and detar the archive every time we want to access the file.

Obviously this does increase disk usage somewhat, but this is less of a
concern for us.